requirements.txt
openlineage-airflow==0.6.2
